pyproject.toml
[project]
name = "abdelrahman-ai-career-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain[google-genai]>=1.2.15",
    "langchain-community>=0.4.1",
    "langgraph>=1.1.4",
    "pageindex>=0.2.8",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "loguru>=0.7.3",
    "langchain-groq>=1.1.2",
    "uvicorn>=0.44.0",
    "fastapi>=0.136.0",
    "gradio>=6.12.0",
]
